Script is updated
