+++ test_scripts/Polices/build_options/014_ATF_P_TC_Timeout_to_wait_response_PTU_HTTP.lua
@@ -1,0 +1,81 @@
+---------------------------------------------------------------------------------------------
+-- Requirements summary:
+-- [PolicyTableUpdate] Timeout to wait a response on PTU
+--
+-- Description:
+-- To define the timeout to wait a response on PTU, Policies manager must refer
+--PTS "module_config" section, key <timeout_after_x_seconds>.
+-- 1. Used preconditions
+-- SDL is built with "-DEXTENDED_POLICY: HTTP" flag
+-- 2. Performed steps
+-- Application is registered.
+-- PTU is requested.
+--
+-- Expected result:
+-- SDL->HMI: SDL.OnStatusUpdate(UPDATE_NEEDED)
+-- SDL->HMI:BC.PolicyUpdate(file, timeout, retry[]) where
+---------------------------------------------------------------------------------------------
+
+--[[ General configuration parameters ]]
+config.deviceMAC = "12ca17b49af2289436f303e0166030a21e525d266e209267433801a8fd4071a0"
+
+--[[ Required Shared libraries ]]
+local commonSteps = require('user_modules/shared_testcases/commonSteps')
+local commonFunctions = require('user_modules/shared_testcases/commonFunctions')
+local testCasesForPolicyTableSnapshot = require('user_modules/shared_testcases/testCasesForPolicyTableSnapshot')
+
+--[[ General Precondition before ATF start ]]
+commonSteps:DeleteLogsFileAndPolicyTable()
+--TODO: Should be removed when issue: "ATF does not stop HB timers by closing session and connection" is fixed
+config.defaultProtocolVersion = 2
+
+--[[ General Settings for configuration ]]
+Test = require('connecttest')
+require('cardinalities')
+require('user_modules/AppTypes')
+
+--[[ Test ]]
+commonFunctions:newTestCasesGroup("Test")
+function Test:TestStep_PTS_Timeout_wait_response_PTU()
+  local hmi_app_id = self.applications[config.application1.registerAppInterfaceParams.appName]
+  local ServerAddress = commonFunctions:read_parameter_from_smart_device_link_ini("ServerAddress")
+
+  local RequestId = self.hmiConnection:SendRequest("SDL.ActivateApp", { appID = self.applications[config.application1.registerAppInterfaceParams.appName]})
+  EXPECT_HMIRESPONSE(RequestId)
+  :Do(function(_,_)
+      local RequestId1 = self.hmiConnection:SendRequest("SDL.GetUserFriendlyMessage", {language = "EN-US", messageCodes = {"DataConsent"}})
+
+      EXPECT_HMIRESPONSE( RequestId1, {result = {code = 0, method = "SDL.GetUserFriendlyMessage"}})
+      :Do(function(_,_)
+
+          self.hmiConnection:SendNotification("SDL.OnAllowSDLFunctionality",
+            {allowed = true, source = "GUI", device = {id = config.deviceMAC, name = ServerAddress, isSDLAllowed = true}})
+
+          testCasesForPolicyTableSnapshot:verify_PTS(true,
+            {config.application1.registerAppInterfaceParams.appID},
+            {config.deviceMAC},
+            {hmi_app_id})
+
+          local timeout_pts = testCasesForPolicyTableSnapshot:get_data_from_PTS("module_config.timeout_after_x_seconds")
+          local timeout_preloaded
+          for i = 1, #testCasesForPolicyTableSnapshot.preloaded_elements do
+            if(testCasesForPolicyTableSnapshot.preloaded_elements[i].name == "module_config.timeout_after_x_seconds") then
+              timeout_preloaded = testCasesForPolicyTableSnapshot.preloaded_elements[i].value
+            end
+          end
+          if ( timeout_pts ~= timeout_preloaded ) then
+            self:FailTestCase("timeout in PTS should be "..timeout_preloaded.."ms, real: "..timeout_pts.."ms")
+          end
+        end)
+    end)
+  EXPECT_HMICALL("BasicCommunication.PolicyUpdate",{})
+end
+
+--[[ Postconditions ]]
+commonFunctions:newTestCasesGroup("Postconditions")
+
+function Test.Postcondition_Stop_SDL()
+  StopSDL()
+end
+
+return Test
